IGNITE-3443 WIP.

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/query/GridCacheQueryBaseMetricsAdapter.java
@@ -84,7 +84,7 @@ public abstract class GridCacheQueryBaseMetricsAdapter implements Externalizable
         if (avgTime > 0)
             return avgTime;
         else {
-            long val = completed.sum();
+            double val = completed.sum();
 
             return val > 0 ? sumTime.sum() / val : 0;
         }
